move all napi re-implimented code into a `js` module
