+++ src/node.rs
@@ -33,7 +33,7 @@ impl From<iroh::rpc_protocol::CounterStats> for CounterStats {
 /// Information about a direct address.
 #[napi]
 #[derive(Debug, Clone)]
-pub struct DirectAddrInfo(iroh::net::magicsock::DirectAddrInfo);
+pub struct DirectAddrInfo(pub(crate) iroh::net::magicsock::DirectAddrInfo);
 
 #[napi]
 impl DirectAddrInfo {
@@ -48,12 +48,6 @@ impl DirectAddrInfo {
         self.0.latency
     }
 
-    /// Get the reported latency, if it exists, in milliseconds
-    #[napi(js_name = "latency")]
-    pub fn latency_js(&self) -> Option<u32> {
-        self.0.latency.map(|l| l.as_millis() as _)
-    }
-
     /// Get the last control message received by this node
     pub fn last_control(&self) -> Option<LatencyAndControlMsg> {
         self.0
@@ -64,27 +58,9 @@ impl DirectAddrInfo {
             })
     }
 
-    /// Get the last control message received by this node
-    #[cfg(feature = "napi")]
-    #[napi(js_name = "lastControl")]
-    pub fn last_control_js(&self) -> Option<JsLatencyAndControlMsg> {
-        self.0
-            .last_control
-            .map(|(latency, control_msg)| JsLatencyAndControlMsg {
-                latency: latency.as_millis() as _,
-                control_msg: control_msg.to_string(),
-            })
-    }
-
     /// Get how long ago the last payload message was received for this node
     pub fn last_payload(&self) -> Option<Duration> {
         self.0.last_payload
-    }
-
-    /// Get how long ago the last payload message was received for this node in milliseconds.
-    #[napi(js_name = "lastPayload")]
-    pub fn last_payload_js(&self) -> Option<u32> {
-        self.0.last_payload.map(|d| d.as_millis() as _)
     }
 }
 
@@ -95,16 +71,6 @@ pub struct LatencyAndControlMsg {
     /// The type of control message, represented as a string
     pub control_msg: String,
     // control_msg: ControlMsg
-}
-
-/// The latency and type of the control message
-#[cfg(feature = "napi")]
-#[napi(constructor, js_name = "LatencyAndControlMsg")]
-pub struct JsLatencyAndControlMsg {
-    /// The latency of the control message, in milliseconds.
-    pub latency: u32,
-    /// The type of control message, represented as a string
-    pub control_msg: String,
 }
 
 // TODO: enable and use for `LatencyAndControlMsg.control_msg` field when iroh core makes this public
@@ -129,39 +95,6 @@ pub struct ConnectionInfo {
     pub last_used: Option<Duration>,
 }
 
-#[cfg(feature = "napi")]
-#[napi(js_name = "ConnectionInfo")]
-#[derive(Debug, Clone)]
-pub struct JsConnectionInfo {
-    /// The public key of the endpoint.
-    public_key: PublicKey,
-    /// Derp url, if available.
-    pub derp_url: Option<String>,
-    /// List of addresses at which this node might be reachable, plus any latency information we
-    /// have about that address and the last time the address was used.
-    addrs: Vec<DirectAddrInfo>,
-    /// The type of connection we have to the peer, either direct or over relay.
-    pub conn_type: JsConnectionType,
-    /// The latency of the `conn_type` (in milliseconds).
-    pub latency: Option<u32>,
-    /// Duration since the last time this peer was used (in milliseconds).
-    pub last_used: Option<u32>,
-}
-
-#[cfg(feature = "napi")]
-#[napi]
-impl JsConnectionInfo {
-    #[napi]
-    pub fn public_key(&self) -> PublicKey {
-        self.public_key.clone()
-    }
-
-    #[napi]
-    pub fn addrs(&self) -> Vec<DirectAddrInfo> {
-        self.addrs.clone()
-    }
-}
-
 impl From<iroh::net::magic_endpoint::ConnectionInfo> for ConnectionInfo {
     fn from(value: iroh::net::magic_endpoint::ConnectionInfo) -> Self {
         ConnectionInfo {
@@ -179,22 +112,18 @@ impl From<iroh::net::magic_endpoint::ConnectionInfo> for ConnectionInfo {
     }
 }
 
-#[cfg(feature = "napi")]
-impl From<iroh::net::magic_endpoint::ConnectionInfo> for JsConnectionInfo {
-    fn from(value: iroh::net::magic_endpoint::ConnectionInfo) -> Self {
-        Self {
-            public_key: value.public_key.into(),
-            derp_url: value.derp_url.map(|url| url.to_string()),
-            addrs: value
-                .addrs
-                .iter()
-                .map(|a| DirectAddrInfo(a.clone()))
-                .collect(),
-            conn_type: value.conn_type.into(),
-            latency: value.latency.map(|l| l.as_millis() as _),
-            last_used: value.last_used.map(|l| l.as_millis() as _),
-        }
-    }
+/// The type of the connection
+#[napi(string_enum)]
+#[derive(Debug)]
+pub enum ConnType {
+    /// Indicates you have a UDP connection.
+    Direct,
+    /// Indicates you have a DERP relay connection.
+    Relay,
+    /// Indicates you have an unverified UDP connection, and a relay connection for backup.
+    Mixed,
+    /// Indicates you have no proof of connection.
+    None,
 }
 
 /// The type of connection we have to the node
@@ -210,97 +139,6 @@ pub enum ConnectionType {
     /// the address works.
     Mixed(String, String),
     /// We have no verified connection to this PublicKey
-    None,
-}
-
-#[cfg(feature = "napi")]
-#[napi(object, js_name = "ConnectionType")]
-#[derive(Debug, Clone)]
-pub struct JsConnectionType {
-    pub typ: ConnType,
-    pub data0: Option<String>,
-    pub data1: Option<String>,
-}
-
-#[cfg(feature = "napi")]
-impl From<iroh::net::magicsock::ConnectionType> for JsConnectionType {
-    fn from(value: iroh::net::magicsock::ConnectionType) -> Self {
-        match value {
-            iroh::net::magicsock::ConnectionType::Direct(addr) => Self {
-                typ: ConnType::Direct,
-                data0: Some(addr.to_string()),
-                data1: None,
-            },
-            iroh::net::magicsock::ConnectionType::Mixed(addr, url) => Self {
-                typ: ConnType::Mixed,
-                data0: Some(addr.to_string()),
-                data1: Some(url.to_string()),
-            },
-            iroh::net::magicsock::ConnectionType::Relay(url) => Self {
-                typ: ConnType::Relay,
-                data0: Some(url.to_string()),
-                data1: None,
-            },
-            iroh::net::magicsock::ConnectionType::None => Self {
-                typ: ConnType::None,
-                data0: None,
-                data1: None,
-            },
-        }
-    }
-}
-
-#[cfg(feature = "napi")]
-#[napi]
-impl JsConnectionType {
-    /// Whether connection is direct, relay, mixed, or none
-    #[napi(js_name = "type")]
-    pub fn typ(&self) -> ConnType {
-        self.typ
-    }
-
-    /// Return the socket address if this is a direct connection
-    #[napi]
-    pub fn as_direct(&self) -> String {
-        match self.typ {
-            ConnType::Direct => self.data0.as_ref().unwrap().clone(),
-            _ => panic!("ConnectionType type is not 'Direct'"),
-        }
-    }
-
-    /// Return the derp url if this is a relay connection
-    #[napi]
-    pub fn as_relay(&self) -> String {
-        match self.typ {
-            ConnType::Relay => self.data0.as_ref().unwrap().clone(),
-            _ => panic!("ConnectionType is not `Relay`"),
-        }
-    }
-
-    /// Return the socket address and DERP url if this is a mixed connection
-    #[napi]
-    pub fn as_mixed(&self) -> ConnectionTypeMixed {
-        match self.typ {
-            ConnType::Mixed => ConnectionTypeMixed {
-                addr: self.data0.as_ref().unwrap().clone(),
-                derp_url: self.data1.as_ref().unwrap().clone(),
-            },
-            _ => panic!("ConnectionType is not `Mixed`"),
-        }
-    }
-}
-
-/// The type of the connection
-#[napi(string_enum)]
-#[derive(Debug)]
-pub enum ConnType {
-    /// Indicates you have a UDP connection.
-    Direct,
-    /// Indicates you have a DERP relay connection.
-    Relay,
-    /// Indicates you have an unverified UDP connection, and a relay connection for backup.
-    Mixed,
-    /// Indicates you have no proof of connection.
     None,
 }
 
@@ -408,7 +246,7 @@ impl IrohNode {
         Ok(node)
     }
 
-    async fn new_inner(
+    pub(crate) async fn new_inner(
         path: PathBuf,
         tokio_rt: Option<tokio::runtime::Runtime>,
     ) -> Result<Self, anyhow::Error> {
@@ -438,15 +276,6 @@ impl IrohNode {
         })
     }
 
-    /// Create a new iroh node. The `path` param should be a directory where we can store or load
-    /// iroh data from a previous session.
-    #[cfg(feature = "napi")]
-    #[napi(factory, js_name = "withPath")]
-    pub async fn new_js(path: String) -> napi::Result<Self> {
-        let res = Self::new_inner(PathBuf::from(path), None).await?;
-        Ok(res)
-    }
-
     /// The string representation of the PublicKey of this node.
     #[napi]
     pub fn node_id(&self) -> String {
@@ -466,14 +295,6 @@ impl IrohNode {
         })
     }
 
-    /// Get statistics of the running node.
-    #[cfg(feature = "napi")]
-    #[napi(js_name = "stats")]
-    pub async fn stats_js(&self) -> Result<HashMap<String, CounterStats>, napi::Error> {
-        let stats = self.sync_client.node.stats().await?;
-        Ok(stats.into_iter().map(|(k, v)| (k, v.into())).collect())
-    }
-
     /// Return `ConnectionInfo`s for each connection we have to another iroh node.
     pub fn connections(&self) -> Result<Vec<ConnectionInfo>, IrohError> {
         block_on(&self.rt(), async {
@@ -489,21 +310,6 @@ impl IrohNode {
                 .map_err(IrohError::connection)?;
             Ok(infos)
         })
-    }
-
-    /// Return `ConnectionInfo`s for each connection we have to another iroh node.
-    #[cfg(feature = "napi")]
-    #[napi(js_name = "connections")]
-    pub async fn connections_js(&self) -> Result<Vec<JsConnectionInfo>, napi::Error> {
-        let infos = self
-            .sync_client
-            .node
-            .connections()
-            .await?
-            .map_ok(|info| info.into())
-            .try_collect::<Vec<_>>()
-            .await?;
-        Ok(infos)
     }
 
     /// Return connection information on the currently running node.
@@ -523,22 +329,6 @@ impl IrohNode {
         })
     }
 
-    /// Return connection information on the currently running node.
-    #[cfg(feature = "napi")]
-    #[napi(js_name = "connectionInfo")]
-    pub async fn connection_info_js(
-        &self,
-        node_id: &PublicKey,
-    ) -> Result<Option<JsConnectionInfo>, napi::Error> {
-        let info = self
-            .sync_client
-            .node
-            .connection_info(node_id.into())
-            .await?;
-
-        Ok(info.map(Into::into))
-    }
-
     /// Get status information about a node
     pub fn status(&self) -> Result<Arc<NodeStatusResponse>, IrohError> {
         block_on(&self.rt(), async {
@@ -549,14 +339,6 @@ impl IrohNode {
                 .map(|n| Arc::new(n.into()))
                 .map_err(IrohError::connection)
         })
-    }
-
-    /// Get status information about a node
-    #[cfg(feature = "napi")]
-    #[napi(js_name = "status")]
-    pub async fn status_js(&self) -> Result<NodeStatusResponse, napi::Error> {
-        let status = self.sync_client.node.status().await.map(Into::into)?;
-        Ok(status)
     }
 }
